Move heading calculation to core 1 (#3)
In order to improve the consistency and reliability of heading calculations, move the work to maintain the heading to the second core on the RP2040, core 1. This required dedicating a I2C bus to each core, and thus affined IC20 to core 1 and I2C1 to core 0. This was solely done for wire length and placement reasons.

There is no longer an option to calibrate the gyro. Instead, the gyro will always be calibrated on start up if there are no offsets set in the config file (that is, all values are 0), otherwise the offsets in the config file are used. gpio16 now indicates whether the gyro is fully initialized and calibrated, being set HIGH when the start up process is done.